+++ balance-api/src/main/java/com/isaachome/balance/repo/CategoryRepo.java
@@ -2,6 +2,6 @@ package com.isaachome.balance.repo;
 
 import com.isaachome.balance.entity.Category;
 
-public interface CategoryRepo extends BaseRepository<Category,Long>{
+public interface CategoryRepo extends BaseRepository<Category,Integer>{
 
 }
